proxy/requirements.txt
flask
pyjwt
psycopg2-binary
confluent-kafka
